requirements.txt
fastapi
uvicorn
pydantic>=2.0
pytest>=7.0
pytest-asyncio>=0.20
joblib
requests
pandas
numpy
supabase
python-dotenv
postgrest-py
lightgbm
aiohttp>=3.8.0
python-multipart 